+++ crates/symbolicator/src/services/symbolication/mod.rs
@@ -1491,7 +1491,57 @@ mod tests {
         let request_id = symbolication.symbolicate_stacktraces(request).unwrap();
         let response = symbolication.get_response(request_id, None).await;
 
-        insta::assert_yaml_snapshot!(response.unwrap());
+        assert_snapshot!(response.unwrap());
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_source_candidates() -> anyhow::Result<()> {
+        let (service, _cache_dir) = setup_service().await;
+        let symbolication = service.symbolication();
+        let (_symsrv, source) = test::symbol_server();
+
+        // its not wasm, but that is the easiest to write tests because of relative
+        // addressing ;-)
+        let modules: Vec<RawObjectInfo> = serde_json::from_str(
+            r#"[
+              {
+                "type":"wasm",
+                "debug_id":"7f883fcd-c553-36d0-a809-b0150f09500b",
+                "code_id":"7f883fcdc55336d0a809b0150f09500b"
+              }
+            ]"#,
+        )?;
+
+        let stacktraces = serde_json::from_str(
+            r#"[
+              {
+                "frames":[
+                  {
+                    "instruction_addr":"0x3880",
+                    "addr_mode":"rel:0"
+                  }
+                ]
+              }
+            ]"#,
+        )?;
+
+        let request = SymbolicateStacktraces {
+            modules: modules.into_iter().map(From::from).collect(),
+            stacktraces,
+            signal: None,
+            origin: StacktraceOrigin::Symbolicate,
+            sources: Arc::new([source]),
+            scope: Default::default(),
+            options: RequestOptions {
+                dif_candidates: true,
+            },
+        };
+
+        let request_id = symbolication.symbolicate_stacktraces(request).unwrap();
+        let response = symbolication.get_response(request_id, None).await;
+
+        assert_snapshot!(response.unwrap());
         Ok(())
     }